requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.25.0
PyGithub==2.3.0
gitpython==3.1.43
redis==5.0.4
arq==0.25.0
httpx==0.27.0
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.20.0
aiofiles==23.2.1
